+++ src/app/consumers/gauge.component.spec.ts
