+++ src/app/helm.module.ts
@@ -6,9 +6,9 @@ import { HttpModule } from '@angular/http';
 import { HelmComponent } from './helm.component';
 import { DashboardComponent } from './dashboard.component';
 
-import { SignalKService } from './signal-k.service';
-import { NetworkSendersService } from './network-senders.service';
-import { GaugeComponent } from './gauge.component';
+import { SignalKService } from './senders/signal-k.service';
+import { NetworkSendersService } from './senders/network-senders.service';
+import { GaugeComponent } from './consumers/gauge.component';
 
 @NgModule({
   declarations: [
