+++ src/app/senders/signal-k.service.spec.ts
