+++ src/app/senders/signal-k.service.ts
